move font preload into preload state so it's reflected in loading bar
